Moved code generator into common library.
