+++ src/DtronixModel.Generator/Ddl/Build.bat
@@ -1,0 +1,1 @@
+"C:\Program Files (x86)\Xsd2Code\Xsd2Code.exe" DdlSchema.xsd /o DdlSchema.cs /cl /n DtronixModel.Generator.Ddl /l CS /pl Net35 /c ObservableCollection /is /db /xa
