+++ src/DtronixModel.Generator/MySqlMwb/MySqlMwbDdlGenerator.cs
@@ -1,14 +1,9 @@
 ﻿using DtronixModeler.Ddl;
 using System;
 using System.Collections.Generic;
-using System.Data.Common;
-using System.Data.SQLite;
 using System.IO;
-using System.Linq;
-using System.Text;
 using DtronixModeler.Generator.MySql;
 using System.IO.Compression;
-using System.Xml.XPath;
 using System.Xml;
 using System.Threading.Tasks;
