+++ src/DtronixModel.sln
@@ -18,6 +18,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DtronixModel.Tools", "Dtron
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DtronixModel.Tools.Tests", "DtronixModel.Tools.Tests\DtronixModel.Tools.Tests.csproj", "{F44DE2F9-9B42-4FF0-82EB-0CFC0EE58769}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DtronixModel.Generator", "DtronixModel.Generator\DtronixModel.Generator.csproj", "{48F0CF19-181B-4645-826B-6BF63B885A05}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -44,6 +46,10 @@ Global
 		{F44DE2F9-9B42-4FF0-82EB-0CFC0EE58769}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F44DE2F9-9B42-4FF0-82EB-0CFC0EE58769}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F44DE2F9-9B42-4FF0-82EB-0CFC0EE58769}.Release|Any CPU.Build.0 = Release|Any CPU
+		{48F0CF19-181B-4645-826B-6BF63B885A05}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{48F0CF19-181B-4645-826B-6BF63B885A05}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{48F0CF19-181B-4645-826B-6BF63B885A05}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{48F0CF19-181B-4645-826B-6BF63B885A05}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
