+++ src/DtronixModel/TableRow.cs
@@ -2,6 +2,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Data.Common;
+using DtronixModel.Attributes;
 
 namespace DtronixModel
 {
