--- src/DtronixModeler/Ddl/Build.bat
@@ -1,1 +1,0 @@
-"C:\Program Files (x86)\Xsd2Code\Xsd2Code.exe" DdlSchema.xsd /o DdlSchema.cs /cl /n DtronixModeler.Ddl /l CS /pl Net35 /c ObservableCollection /is /db /xa
