+++ src/DtronixModeler/DtronixModeler.csproj
@@ -19,7 +19,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="System.Data.SQLite.Core" Version="1.0.113.1" />
     <PackageReference Include="Tools.InnoSetup" Version="6.0.5" GeneratePathProperty="true" />
   </ItemGroup>
