+++ src/DtronixModeler/Xaml/AboutWindow.xaml.cs
@@ -4,6 +4,7 @@ using System.Diagnostics;
 using System.IO;
 using System.Reflection;
 using System.Windows;
+using DtronixModel.Attributes;
 
 namespace DtronixModeler.Xaml
 {
